Fix for flattener refs being strings or tuples (#155)
The flattener emits pseudo-JSON schema, but with $refs as tuples.
This is bad when the sub-schema is walked twice, e.g. when
flattening combiners (first via _flatten_combiners + _walk in
_flatten_object_type, and second via _walk in _flatten_object_type),
where _flatten_ref_type is expecting a string $ref.

One solution is to encode refs when they are returned by the
flattener, so they are always strings, and then decode them when used.
But that's more encoding and decoding operations.

The solution I've chosen here is to make _flatten_ref_type aware of
already processed refs, which is a quick fix, but one that feels
not quite right, because of the pseudo-JSON-schema-ness.