$ref for draft-07 changed to HTTPS endpoints (#181)
Starting to build out validation logic in handlers and hit some issues with certain
libraries trying to hit HTTP endpoints. That can be fixed in the code, but since
there's a valid HTTPS endpoint for the JSON Schema it made sense to flip to that
as this also fixes the problem and avoids others having same issues later.