Adding handlers section to schema (#213)
* Adding VSCode workspace file to .gitignore

* Fixing sourceUrl to new organization

* Adding handlers section to schema

Handler schema needed to be added back to the resource spec to provide some registration metadata as well as declarative permissions to operate handlers for a particular resource type. Handlers are optional, but if supplied, all must be defined. RPDK may need additional changes in the future around this, but I want to put the baseline schema back in place.

* Uplifted protcolVersion, runtimeType definition