Add support for update overrides (#309)
* Add support for update overrides

Formerly, we generated update inputs and used the CREATE override for
fields that needed to be overridden. The issue with this is that it's
not possible to test update handlers with those fields to be updated.
Now people can specify fields to override with upon update, falling back
to the creation overrides if an UPDATE override is not provided.