Fix flaky unit test (#637)
The unit test in question was depending on a specific folder to exist. There are other tests which will create that folder, but I suspect the tests are executed in a deterministic order. Fixed this by changing the test to use a shared method which includes the creation of said folder.