[Modules] Support parsing Yaml fragments with CFN short forms (#699)
* Issue 693 [Modules] Support parsing Yaml fragments with CFN short forms

* Issue 689 Only use cfn-flip for parsing module fragments

* Issue 689 Apply cfn-lint to the original file

To prevent any false warnings from we should apply cfn-lint to the original file and not rely on parsing and then serializing to json before running lint.

* Issue 693 Better error message when there is no module fragment file

* Issue 693 Ensure unhandled exceptions in cfn-lint don't block the user

When something unexpected happens when cfn-lint parses the template we don't want to block the user from running cfn submit. This should be an edge case though as we already parse the template before running cfn-lint albeit using a different library.