Issue 640 Ignore unrelated files when creating modules zip (#700)
When we validate the fragments folder we only look at files ending on json or yaml, but when we run cfn submit we just zip up everything in the fragments folder. This can cause the module registration to fail when there are unrelated files in that folder (such as files created by an IDE for example). This PR changes the file-zipping logic to only include the relevant fragment files.