Add Community Section. (#954)
To continue to drive awareness to our CloudFormation Discord and give the community another place to ask questions we are adding a community section to our popular repositories. 

I placed this at the bottom of the ReadMe but open to suggestions on where in the ReadMe this section should be placed.